refactored totalbudget. made components to display total budget and expenses.
